+++ ocs2_oc/include/ocs2_oc/rollout/Rollout_Settings.h
@@ -151,7 +151,7 @@ inline void Rollout_Settings::loadSettings(const std::string& filename, const st
 	}
 
 	try	{
-		maxNumStepsPerSecond_ = pt.get<double>(fieldName + ".maxNumStepsPerSecond");
+		maxNumStepsPerSecond_ = pt.get<size_t>(fieldName + ".maxNumStepsPerSecond");
 		if (verbose) {  std::cerr << " #### Option loader : option 'maxNumStepsPerSecond' ................ " << maxNumStepsPerSecond_ << std::endl;
 		}
 	}
